fix(examples):修复基准测试和并发控制示例中的类型错误
- 将 `regex_str` 改为借用以避免所有权转移问题-修复任务 ID 类型不匹配问题，统一转换为 u64 类型

+++ action_dispatch/examples/concurrency_control.rs
@@ -38,7 +38,7 @@ fn benchmark(mode: &str, single_thread: bool, task_count: usize) {
         .map(|i| {
             thread::spawn(move || {
                 let task = Task {
-                    id: i,
+                    id: i as u64,
                     name: format!("任务{}", i),
                 };
